Get rid of inheritance hierarchy
The main reasons are simplicity, ease of use, and concentration on the
essentials. While the inheritance hierarchy was conceptually appealing, the
implementation was rather intricate, let to some redundancies, and had a
(slightly) negative impact on performance. The utility of the edge set was
doubtful. Point sets can still be represented using the SurfaceMesh class.

This commit

- moves all functionality back to SurfaceMesh
- removes GeometryObject, PointSet, EdgeSet, and associated classes
- renames BaseHandle to Handle