Changed some configuration checks to allow MinGW to build
Using the MinGW64 gcc compiler built-in to CLion, some errors are reported because cmake adds /wd4700 and it's friends to the command arguments, but gcc doesn't understand them, and reports them as linking errors.  Changing the IFs to check for the MSVC compiler, instead of windows, these flags will be added only when they will be understood.

Added an IF for win32 to add some compiler definitions that apply to MinGW as well.

Signed-off-by: Chris Samuelson <chris.sam55@gmail.com>