Implement new_face( f_to_copy_properties )
This new version of new_face will propagate the properties from the
argument f_to_copy_properties to the newly created face.

An alternate approach would be to instead pass an invalid Face as a
default argument and testing for that argument could make this optional.
I.e.

Face new_face(Face f_to_copy_properties = PMP_MAX_INDEX )
{
...
if ( f_to_copy_properties < PMP_MAX_INDEX )
{
// do copy
}
...
}


However, this kind of use of default arguments does not seem prevalent in
PMP, so I went with a separate method instead.


The same kind of thing could be done for vertices and edges, but I'm
concerned with faces for now.