Split up eigen matrix multiplication to ensure specialised algorithm used
On an mvsc (visual studio 2022) build of pmp, when doing some profiling of remeshing I found the line "L = -0.5 * D * L" was taking 2.5 minutes for a 400,000 vertex mesh. 

Rearranging the statement to move the "-0.5 * " to the line above caused the statement to only take ~25 ms.

My hypothesis is that this rearrangement caused eigen to use an optimised templated specialisation for Diagonal * Sparse, which it was previously missing.